Fix some absent gems and a crystalized boss.
grunt_lair_end_beastmaster lost the titular troll when the earthy gem was
added to that vault. Also, all the scaling vaults for Depths that can have
a Zot entrance weren't touched (same with the D vaults that could have a
Depths entrance), so they didn't place any gems properly. 866d48a was an
admirable effort, nonetheless, but this should fix up more of its wake.

Also did some cleaning up while I was looking through all of these. Of
minor note was finding out grunt_runaround was set up for D:15 but never
actually placed. I'm heavily nerfing the Depths entrance it set up, though.
D:15 is not a place for a titan, two hell knight bands, and a shrike in the
same room.

Eventually, I'd like to adjust most of these vaults to place the gem not
underneath the rune or other loot or on the Depths / Zot entrance, so that
they can be admired on the floor as the sparkly gems that they are.
Probably later, though.