Re-order alligator attacks
As with trample, dragging needs to be the last attack in a monster's list
of attacks, or it won't do further attacks after it moves. It's a little
weird for them to tail-slap before biting, but I wasn't intending for
alligators to lose quite so much melee damage against higher AC characters,
so it has to work out this way unless a fair ways more material is
rewritten.