Fix a sneaky inugami bug
We didn't erase CANINE_FAMILIAR_MID when we changed levels, meaning
that we might accidentally believe that some other random monster on
the new level was our beloved dog. When we cast call canine familiar,
we could end up buffing and healing a random, probably hostile monster,
or just being unable to summon a new dog. Pretty good!

This whole mid-based approach is very scary to me (it feels quite
fragile), but it's what we're doing for spectral weapon, so maybe it's
fine? I wonder about tagging the dog instead of the summoner, though.