buff ice ziggurats (#3401)
[Committer's notes: Simulacra in ziggurat previously called directly
from HD versus the ziggurat's abs/depth, which would end up getting
stuck later on shrikes and two of the giants. The specified list for
simulacra is a good idea, but it is too much of a malevolent meme
for zig dippers to possible instantly explode against a 20-headed
hydra simulacrum. Used a different specific list instead. Also,
wendigo do in fact overlap too much with Cocytus. Rather than pull
further from Cocytus, use ice monsters that don't- ironbound
frosthearts, and walking frostbound tomes. Both are probably still
fine for Ziggurat dippers, and one of the layouts places the start
of the floor by a wall.

Closes #3401.]