Remove undocumented Chei/gem interaction
Was surprising and probably not needed for Chei to slow gem timers.
We can reconsider this later, of course, but if we do we should
probably make Chei slow the rate at which gem_time_spent accumulates,
so players aren't incentivized to consider swapping to Chei while
almost out of gem time.