Revamp Sticky Flame (the status and the spell!), give to Alchemists
Sticky Flame, the player spell, has been widely considered to be a
little sad for a while; the attack of opportunity changes hurt kiting
with it a lot, and there are generally more appealing fire options at a
similar level. And Sticky Flame, the status used *against* the player,
has been even sadder for far, far longer. It doesn't scale with the
enemy applying it and does such a trivial amount of damage that it can
be completely ignored in virtually all situations.

This commit aims to address both of these issues simultaneously and give
Alchemist a new toy at the same time (as well as a better tool in their
kit to deal with rPois enemies).

Sticky Flame (the status) now does more damage per turn than it used to
(*far more* against players) and has a generally longer duration, but is
removed rapidly when the victim moves instead of performing other
actions - essentially a sort of 'reverse barbs'. The per-turn damage of
the player status also scales directly with enemy spellpower, so that
both smoke demons and balrugs can threaten appropriately for their
depth. The idea is that this status will do significant fire damage over
time unless the player actively attempts to put it out - but not so much
that it is always correct to drop what you're doing and do so.

The status effect still does fixed damage per turn to monsters (though
higher than before), but monsters *also* put the status out much faster
when they move - meaning that when the player lights a monster on fire,
they will now want to stand in place next to it rather than move away and
waste most of their spell, which should hopefully bring interesting
positional decisions, especially when combined with the rest of alchemist's
starting kit.

Sticky Flame has moved from Conjurations/Fire to Alchemy/Fire and added
to Alchemist's starting spells. It should do meaningfully more (and more
consistent) damage to targets, so long as you can actually remain near
them.

Also, the monster spell name 'Sticky Flame Ranged' was always a
historical holdover and somewhat confusing considering the multiple ways
in which it was different than Sticky Flame, so it is now further
divorced from the player spell and renamed Pyre Arrow. It deal no impact
damage at all (while the player spell still does, as a means of
additional scaling, considering that the damage of sticky flame against
monsters is fixed). Monster spell info displays the *per turn* damage
taken from the sticky flame status this spell applies. (Alas, this is
currently not true and also complicated for the player version when used
by player ghosts...)

I have also done a bunch of internal renaming and light refactoring. No
longer do we have a DUR_LIQUID_FLAMES, but an ENCH_STICKY_FLAME, and
then apply them both with a function called 'napalm', but instead use
sticky_flame consistently.

There are a LOT of numbers and formula changes in this commit, and I
have done my best to make them vaguely reasonable, but I have no doubt
that some will need playtesting adjustment - especially since this is a
new dynamic being introduced both for and against players.