Poisonous Vapours -> Mercury Vapours
Poisonous Vapours was in a somewhat awkward position, where at very low
power, it would often inflict less total damage than sting for twice the
mp - something a very early character often cannot afford, even with
vapour's more generous targeting. It became much better with more mp and
spellpower, but by that point you could often cast OTR which usually
outclasses it.

This attempts to give it secondary appeal, even while sandwiched between
those two starter spells, and give a little bit more alchemy flavor.

It is still a smite-targeted direct poison spell, but it additionally has
an HD-based chance to inflict weakness on both the target and any adjacent
to it - regardless of poison resistance. It is also slightly more generous
with posion application if the target was not already poisoned, and a
little less generous at reaching full stacks. I also reduced its range to
6.

The current weakness success chance is:
100 - max(0, (victim_hd * 12 - pow * 2 - 10))
which aims to make it fairly reliable against early enemies, but drop off
severely against later ones - even at max power. (Enemies not in the center
of the spell have only 2/3rd the normal chance to be affected - I am not
sure if this is a necessary complexity or not).