Don't heal poly'd mons (MysticSailboat)
Polymorphed monsters getting a huge heal was added in 0c496e0be39
(0.4, 2008) without explanation. It doesn't seem intuitive or needed,
and is actively counterproductive for Nemelex's Degeneration card
in combination with other Destruction cards, so let's cut it.

(It was also completely undocumented, as far as I can tell. Of course!)

Shapeshifters get to keep healing on each shift, as before. Slimified
monsters explicitly get a full heal, since that feels right to me
( ✨ vibes ✨ ).