New monster: weeping skulls, for Ossuaries / Abyss ( / Crypt vaults)
Ossuaries are in a weird place relative to the rest of early D, though not
nearly as plain as Sewers were prior to ac1dc8b and 5e5e084. Plain mummies
and derived undead are both, strictly speaking, a unique monster and a much
later monster type for the Ossuary default set. The problem is that without
the interference of traps (heavily reduced over time) or very large numbers
and no chokepoints, there's a lot of boring kiting and picking-apart the
slow zombie hordes for a variety of different characters. The energy rework
of attacks-of-opportunity help them a little bit, but not enough to make
zombified versions of D:2 + human and elf zombies particularly plus slow
mummies very interesting. There's a few ossuaries with unique gimmicks,
like the wraith ones or the open layout ambushes or silence or scorpions
or bog mummies, but overall, the rest are very, very plain.

Weeping skulls are a new monster meant to be occasional Ossuary bosses as a
noticeable spike of danger beyond the current sad state of normal mummies,
alongside another future enemy type. Somewhat frail undead, they use the
newly-altered clouds of excruciating misery by randomly and sparsely
spawning them in random squares of a two-tile ring around themselves (but
never directly on top of the player). They also have a variant of the
cloud-line breath catoblepae and ushabti have, spawning brief misery clouds
while skipping past those that would be vulnerable (non-Undead). These both
make running from zombies and mummies both include a fair bit of deliberation
over walking around clouds or taking small amounts of unavoidable damage to
walk through them, which should add a fair bit of interest to many
gimmickless and weaker ossuaries all being roughly the same.

They also show up in the Abyss as natural spawns, once more taking off a
little bit of the very large weight small abominations have early in the
Abyss, and spawning with a second skull when any deeper than Abyss:1. Most
enemies in the Abyss will be immune to the clouds, and they can sometimes
slow the player retreating from far more dangerous enemies around, which
is a cute use for weaker monsters there.

Their tile is by Sastreii.