Add Bombardier Beetles to help tutorialize new sticky flame
The changes to monster sticky flame make it substantially more dangerous
than it used to be, and players used to the old status quo may not realize
that moving will greatly reduce damage taken. I made sure to include this
in pyre arrow's description, but why would a player who thinks they know
how it works look at it again? (And it's too intrusive to tell the
player every turn "Are you sure you want to stay still?")

So I am adding a monster with pyre arrow to early D. It is intended to be
scaled so that it isn't terribly likely to *kill* players, but may still
naturally teach them how sticky flame works before they reach enemies who
*can* kill them with it.

It is a speed 6 creature, so it gives the player plenty of time to run away
from it while on fire (or even incidentally discover that they can shake
off sticky flame via movement when they charge *towards* it.) and is
otherwise fairly harmless in melee.