Revamp Sun Moths
Sun Moths are... weird. Why are they holy (in a branch with multiple
notable natural necromancy things, for that matter)? What is the halo
supposed to accomplish? Why do they resist electricity?

They've effectively just been irresistable bolt monsters with decorative
sticky flame and while that isn't *nothing*, said bolt of energy is slated
to get a will-lowering effect this version and that *also* doesn't do
anything in Spider, so this seemed a good time to give them a shake-up.

They now have a higher intensity pyre arrow and lose energy bolt and
AF_STICKY_FLAME to pick up AF_SWOOP. They will light you on fire and
intercept you when you attempt to shrug it off. They lose rElec, become
rF++ instead of rF+, and become a little more fragile. I'm perhaps
needlessly worried about how dangerous new sticky flame is since I don't
have good data on that yet, but the *idea* is that these become one of the
most notable sources of it in the game, since it's their primary means of
threat (whereas it is secondary on most other enemies that have it).

I have also rewritten their description. I am rather fond of it, tbh.