fix: correctly show email in change email ui on webtiles
Luckily, this bug was entirely cosmetic; email was not actually set to
null and it was even updated correctly. But it prevented both logging
the new email, and showing the new email (or old email) to the player in
the webtiles UI.

This fix will require a webtiles server restart to activate.