Tweak Mercury Vapours in a couple of ways
The range is greatly reduced (6->3), weakness formula is tweaked (generally
lower chance at most levels, and especially at higher enemy HD), and it now
won't inflict any poison on creatures with rPois.

In turns out that the rules about whether sources can apply poison to rPos
monster 1/3 of the time are very inconsistent, but there's a *mild* lean
towards this being true of physical delivery sources (eg: venom brand) and
BEAM_POISON (which the player largely has no access to) and false for
several other spells - notably OTR in the same starter book. Since Sting
already is notable for being part-physical, I think it's fine to make rPois
enemies fully immune to the poisoning of this spell.

But I do want to keep the weakness effect that operates on a different
vector and have slightly tweaked the description to maybe make that
clearer.

Finally, I have allowed players to cast it on themselves, since this was
requested multiple times (believe it or not). By the same rules as enemies,
rPois will protect you from the poison application (but not weakness!)