Blastmotes: explode on wait (elliptic, Implojin)
Per discussion, it seemed simpler and more intuitive for blastmotes
to explode when a creature (incl. the caster) ends their turn in the
blastmotes, rather than only when they move in. The motes bump into
you, etc. (They're still sleepy when you cast them, so you get a turn
of leeway to bail.)

We could cut the special logic to explode blastmotes on move from
monster.cc and player.cc, but the results were less visually pleasing,
so I left them in.

We might want to add warnings for waiting in motes, but eh - it's a
mistake you won't make too many times. :)