Greatly increase bombardier beetle's xp
Due to how the game automatically determines this, it was far lower than
other monsters of comparable threat. (I wonder if it might even have caused
the UI to suggest to players was trivial when it wasn't?)

Still keeping a close eye on these little guys' lethality and may nudge
them downward, depending on how it shakes out.