Guardian Golem -> Blazeheart Golem (lv 4 Summon/Fire)
Guardian Golem has had an identity crisis for a while. Nominally about
protecting your other summons, its most effective usage was actually as a
bomb - with injury bond more commonly serving as a means to help it blow up
reliably than to actually protect anything. It also did wildly high damage
for a level 3 spell - a whopping 3d20 at 0 power, plus additional flame
clouds.

But the thing is - the bomb playstyle was actually kind of fun! There were
positional considerations with getting the bomb to blow up on a timely
basis and not also engulfing yourself in it and it felt like it had obvious
use-cases that weren't shared with other spells. Having summoning spells
that feel distinct and have distinct uses is good, and so I aim to retain
that here, while splitting out the ally-shielding part of the spell and
(hopefully) making its damage more reasonable for its level.

Summon Blazeheart Golem is a level 4 Summoning/Fire spell. The new golem
has a melee attack (and a pretty good one, at that!) but a unique property
where the flame inside it can only be sustained if the summoner stays close
to it. Outside of a small grace period to make it more fluid to use when
the both of you are moving, the golem will go completely dormant so long
as the summoner is non-adjacent to it.

The golem's attack is slow, has AF_FIRE, but also damages the golem for
25% of its max hp whenever it lands a blow. When the golem dies, so long as
it is not dormant, it leaves its volatile core behind. This core will
explode in a 3x3 fire explosion (with clouds) after the next player action.
The timing is handled so that the player should ALWAYS get a chance for
exactly one action between seeing the core appear and it exploding, no
matter how fast or slow you are. A dormant golem causes no explosion at
all.

The idea here is that you can still use this golem as a tool to cause loud
fiery explosions, involving unique positional play. The summoner has to be
willing to step into the fray with the golem, and then plan to get out
again. You cannot throw it at problems from a safe distance (but also are
not randomly punished for engaging with it). In fact, the idea is that it
should be competitive with Lightning Spire in the starter book, but for
different situations/playstyles - Blazeheart Golem wants you to be in the
fray with it, while Lighting Spire supports a more evasive/mobile approach.

All damage numbers are still provisional and will be tweaked based on
feedback, but for reference the explosion currently does between
3d11 - 3d20 based on spellpower (compared to old golem's fixed 3d20).