fix: Better placement of the Crypt gem (dilly)
Due to how the shared Crypt loot randomization works, the ivory gem
could end up in some obscure places on e.g. evilmike_haunted_forest,
namely locked away with unsightly ghouls behind iron grates. This commit
moves the gem off of one of the special loot piles that gets randomized
to various possible locations in each map, and onto its own glyph. This
glyph now appears in a central location at the "end" of every Crypt:3
end vault. On a couple maps the gem replaces a single gold pile, which
is no great loss to the player, and on the rest no loot items are
sacrificed.