Add Martyr's Knell (level 4 Summon/Necromancy)
The *other* half of Guardian Golem.

Shielding allies via injury bond is a decent concept, but ironically, pure
summoner builds have less use for this than necromancers. Summons are more
easily replaceable, often fewer in number, and if we look at the starter
book in particular, there are few targets worth protecting - spammals and
imp are too low level to matter, inugami already offers a mechanism to
protect it, and lightning spire is often out of enemy range. Zombie armies
on the other hand....

Martyr's Knell summons a Martyed Shade, which is a fast-regenerating, but
non-attacking ghost that offers passive injury bond, just like guardian
golem used to. But if it dies before expiring, the pain of re-experiencing
its own death will transform it into a flayed ghost (and instantly flay a
couple random hostile targets). A flayed ghost can be powerful, but since
its damage is so transient (and it is not especially durable), either
positionally defending it or quickly killing its targets is important if
you want it to matter.

The hope is there are a number of ways to use this spell. It can defend an
undead horde and engage more safely from behind it by dying quickly, or it
can be used with limited other summons (or even on its own, with deliberate
positioning) as an offensive tool that requires the player to ensure it
doesn't die early.

Spellpower affects both shade and flayed ghost hp. The latter helps make
the flaying stick around a little better. The former may not be *strictly*
upside, but it seems reasonable to make it just a little harder to get your
more powerful ghosts (and it makes it better at tanking, of course).

There is also a working enemy implementation of this spell, and I think
there is definite design room for an enemy unique using it. (And, of
course, there's always Mara).