Simulacrum -> Sculpt Simulacrum (6 Ice/Alchemy)
Current simulacrum is by no means a bad spell in isolation, but it has
the misfortune of being packed in among multiple other necromancy ally
creation options that are either easier to cast, more powerful, or
arguably both. But simulacra are cool, so let's try opening things up by
moving it out of the necromancy school entirely, where it may face less
competition in its role.

To differentiate it further from corpse-based necromantic means of
generating allies, the spell is redesigned. It is now a range 1 touch
spell that creates 2-3 blocks of ice nearby, which will transform into
simulcra of the touched target after 3-5 turns (plus an extra 1-3 turns
for each subsequent block of ce). This can be cast repeatedly on the
same enemy, with no limit on the number of ice blocks which can exist at
once, but there is a cap of 5 active simulacra and a new one awakening
will expire your oldest. These simulacra are also shorter-lived than the
previous version of this spell.

This creates a dynamic where you often want to deliberately get in
threat range of more dangerous enemies, but still have to wait for a
payoff, which should hopefully feel a little different from other ally
spells. It is worth noting that the player also cannot swap through ice
blocks, so this can create dynamic terrain problems - and creating many
in a row may fence you in with whatever monster you considered dangerous
enough to want to copy.

It is worth noting that there are a lot of balance knobs to tweak here -
durability of ice blocks, number created at once, delay before summoning
things out of them, etc. The original concept of the spell even
inflicted freezing damage with the initial touch (and generated fewer
simulacra overall), but it was suggested that it might be better to
focus on just one aspect here. So let's try it out!

(Also, since there was poor visual clarity between blocks of ice and
simulacra in tiles, I tried to differentiate the block of ice tile's color
from simulacra as much as I could while still looking like ice, as well as
limit it to only a single shape, for a consistent silhouette. I think this
helps at least a tiny bit?)