Agony -> Curse of Agony (for players)
Agony has been widely considered an unimpressive spell in player hands for
quite a while. In earlier parts of the game, its damage often compares
unimpressively to Vampiric Draining against many enemies - somewhat sad
when these are both starter spells and Agony is 2 levels higher. It can
find a niche later in the game against some enemies, but ideally the
capstone of any player background is something you are excited to get
online. Ranged Agony was considered problematic in past, so let's try a
different approach.

Curse of Agony is a range 3 hex that inflicts a brief curse on the affected
enemy. The next two times that enemy is struck in melee by the caster, they
will lose half their current health - AFTER the damage from the melee
attack applies (so somewhat less than 50%, in practice). The small bit of
range lets the player try to overcome an enemy's Will *before* they're in
range of Vampiric Draining or melee, but still requires them to engage
directly instead of using zombies to trigger the agony effect.

Its power multiplier is cut in half compared to Agony, which should make it
somewhat less useful as the game goes on, but ideally more useful in
earlier parts of the game than Agony itself was.

This is also a chance to make some monster spells less confusing. Old
'Agony Range' is now just Agony again, while Fenstrider Witches get
Agonizing Touch. There is no functional change, but I think it should be
a little more obvious to new players that Fenstriders can only do this in
melee range - especially now there there is no player analogue to this.

Curse of Agony is replaced with Agony on player ghosts / illusions, so they
have something to cast in that slot.