Let Fugue of the Fallen work with Kiku wretches, tweak Kiku gifts
After some discussion, it seemed there was wide agreement that it was fine
if the deaths of wretches gave stacks for Fugue. Importantly, this lets
there be a second spell that benefits from wretches between levels 2-5, so
Kiku can guarantee something less sad than Necrotise. Reliably turning
wretches into slaying is powerful, but the piety cost of calling wretches
multiple times to do so should still be meaningful.

Kiku now will always give one of Animate Death or Fugue of the Fallen in
their first gift. Martyr's Knell is added to the second gift list,
replacing Sculpt Simulacrum (since it is no longer necromancy).

This has the potentially interesting consequence that there is a 50/50
chance of your early necromancy gifts pushing you in an ally vs. melee
direction - and in fact, that Kiku can gift a spell several starts might
want that *isn't* in the Necromancer starter kit.