Fugue tweaks
Prevent capstone pain damage from hitting firewood.

Make the 2d5 capstone pain damage scale with spellpower, starting lower at
2d3 (up to 2d7 at 100 power), partially to compensate for how much easier
it is to trigger the capstone at low skill via Kiku wretches. Perhaps this
complication is unnecessary? (Or perhaps it would make more sense to
trigger pain on only some subset of hits and make it hurt more when it
does, for less message spam?)

Add the ambient noise-making message from Fugue to default runrest ignore,
so that it isn't so obnoxious to try and rest after a fight due to all the
wailing.