Fix several Battlesphere targeting bugs, refactor, tweak slightly
Battlesphere triggering was originally split into two functions:
aim_battlesphere(), called immediately before the trigger spell fires, and
trigger_battlesphere(), called immediately afterward. This was so that it
could save the position of the player's firing target, in case that target
died to the triggering spell. Since the changes several years ago to make
battlesphere fire at a RANDOM target instead, this split was needless code
complexity - we can simply pick a target AFTER the spell fires.

In fact, doing so fixes a bug that could frequently cause the battlesphere
to appear not to trigger for no obvious reason: if the random target it
picked before the player spellcast was itself killed by the player
spellcast. This was actually very common with AoE spells or against only
two visible targets (a 50% chance to skip its turn whenever you kill one!)

Battlesphere would also silently fail if the random target it picked was in
line of sight of the player, but NOT the battlesphere itself - another
situation relatively common when corners were combined with the
battlesphere's idle movement. Now it considers only the player's los when
picking a random target.

Finally, after a target was picked, if there was no safe line of fire to
it from the battlesphere's current position, it would attempt to find
another nearby spot and shoot from there. But the search radius was smaller
than how much ground it could cover in a turn, so it would sometimes
silently fail while appearing to have a valid firing angle. It is still
possible to fail in this situation, if the battlesphere ends up more than a
full turn away from ANY viable firing position (often because the player is
body-blocking most of them), but it still meaningfully improved.

Additionally, I have tweaked the target select to prefer shooting at the
same target the battlesphere shot at last turn - so that it essentially
'locks on' to a given enemy - even if the player has no control over which.
It felt far too common for the battlesphere to spread its shots around
uselessly, doing insignificant damage to multiple targets - now it should
be a little more dogged about taking SOMETHING down.