Make Spellforged Servitor a little less gameable, nerf mildly
The fact that Spellforge Servitor could inherit a player spell that had a
whopping ~50% fail rate meant that it was often correct to learn a spell
that you could never viably cast. And since servitor spellpower is based on
the power of the summon, rather than the spell it's casting, this barely
mattered. Plasma Beam was the most common offender, since it doesn't even
share a single school with Servitor.

Now the player needs at least a 20% success rate in a spell for servitor to
cast it. That's still a pretty high fail to use on a conjuration, but not
outside the realm of someone trying it.

This also nudges servitor's HD in player hands down just slightly. It
should still be a very strong spell.

(Also, I moved LRD above Stone Arrow in the spell priority list, since I
think that normally the player would prefer it cast that AND it's higher
level so it's probably not unintuitive that it might?)