Match ally speed outside combat (elliptic)
As a UI convenience, make allies speed up to match the player's
move speed if there are no enemies nearby. This is a common minor
pain point for players (needing to hit ctrl-t to match travel speed),
and there's no reason players should need to worry about it.

This should hopefully be mostly transparent and unnoticeable for
players.