Seriously update docs/develop/spells.txt
This is an immensely old document, referring to many removed schools and
removed spells, mostly unchanged from seven years ago in f9017aa. Maybe
it's mostly unused, but it's still sitting around in the source.

I'm giving it a pretty quick first draft update. Almost anything is better
than still referring to Charms or demon-summoning or the time when like
66% of Earth spells were Transmutations. At some point, maybe one of the
devs who originally worked on positional magic can update this further with
more precise philosophy, but it's at least a document actually attached to
Crawl compared to version-specific Github pages.