Add Sastreii's new Depths tiles
A branch-specific version of stone for all of Depths, to which I added a
quick plainer variant to reduce a bit of noise when a lot of stone walls
are visible at once, as well as new ground floor tiles and dramatic
columns for statue-replacements, now employed in most of the ever-iconic
Depths encompass vaults. The floor tiles and columns hould be reasonable
for other vaults, old and updated or new, to specify or leave out as
their theme fits. It should be used sparingly for particularly dangerous
and large vaults, though.

Also, to try to help with branch appearance differentiation, don't use
these tiles or columns outside of the context of Depths itself. (Maybe
sometimes the Depths entrance in D would be fine?)

(This also came with a little bit of vault clean-up. Should do some
more for the sake of gems and vault header standardization eventually...)