More heavily nerf Blazeheart Golem
The golem seems to be pretty popular! It also seems to be too strong. Since
the exploding part of it is the most mechanically flashy, I've aimed most
of this nerf at its melee attack.

It loses AF_FIRE entirely. It gains a bit of base attack damage in
exchange, but this should still be a significant reduction in damage. Its
attack now also removes 1/3 of its hp on impact, instead of 1/4, meaning it
will fall apart more often and also need to be recast more often for the
same damage output. I've also reduced its AC a little (though I'm not
really sure much this matters in either direction)

Let's see how it feels after this.