Don't lungepage recklessly (Hal)
If you rampage toward a monster and then kill it with lunge, don't
continue by moving into its tile. Even if we prevented moving into
deep water or lava, this could still result in the player moving
into traps, dangerous clouds, or exclusions, all of which seem like
potentially nasty surprises. Let's settle on a destination tile when
we start the move and then stick to it.