Reduce charmed / bribed allies' reckless disregard for collateral damage
There's some old, odd code that reduces the foe_ratio value for monsters
casting spells- basically, how worthwhile it is to hit allies versus hit
foes, by evaluating them against each other with collective HDs and XLs-
if the caster has an only temporary attitude change. This was originally
meant as flavour and balance for both Summon Greater Demon (where
successful summons were hostile but temporarily charmed) and Control
Undead (which was never permanent)- if they already will become hostile
eventually, then they could hypothetically be fine with "accidentally"
hitting the player. Since we have neither of these any more, it is instead
a strangely opaque component of wands of charming and Gozag bribes that
those affected will ever-so-helpfully hit almost any number of players
and allies as collateral damage. This is unideal behaviour for being
completely unexpected as a downside to either, so this special-casing
has been cut.

TODO: There's some interest in further reducing the capacity for allies of
the player to hit the player- see 98ec557 and 98e7288. This would need a
significant undertaking to rewrite the current use of foe_ratio due to it
being set constantly elsewhere and also being used for hostiles versus the
player as well as allies alongside the player. Removing it entirely without
a lot more supporting code makes either monsters be willing to hit as many
others as possible if this also hits the player, or it'd be possible to
stand by a rat at all times to prevent any spells being cast. Either way,
it's still quite fundamental for how fire giants, de sorcerers / high
priests, and dr scorchers don't fireball themselves. Probably this could
all do with some preliminary work, but not so close to 0.31's release.